software/ctd-collector/requirements.txt
PyYAML
atlas-i2c
pony